refactor(proxy): modularize functionality better
